requirements.txt
anyio==3.6.2
APScheduler==3.10.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
joblib==1.2.0
nltk==3.8.1
praw==7.6.1
prawcore==2.3.0
python-telegram-bot==20.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
update-checker==0.18.0
urllib3==1.26.14
websocket-client==1.5.1
